EY-4962: Rename dao
